requirements.txt
asgiref==3.3.1
certifi==2019.9.11
Django==3.1.6
django-3-jet==1.0.8
django-jet==1.0.8
djangorestframework==3.12.2
Pillow==8.1.0
pipenv==2018.11.26
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
virtualenv==16.7.7
virtualenv-clone==0.5.3

django-filter~=2.4.0